Show Token Art: Reference the specific selected token
The Show Token Art previously referenced the prototype token which means when you changed a specific token and used Show Token Art then it would show the old prototype token art. 

This also means that if you assigned a wildcard image at the prototype level, the original function would return an error of image not found. Whereas referencing the specific token means that you'll get the specific wildcard art that is selected.